Fluxo edit payment

+++ Asaas/grails-app/domain/asaas/Payment.groovy
@@ -8,7 +8,7 @@ import asaas.Payer
 
 class Payment extends BaseEntity {
   
-  Integer value
+  BigDecimal value
 
   Date dueDate
